Explorer Selector

Use the explorer module to select a node.

explorer_selector_ui(id)

explorer_selector(input, output, session, .values, .root_node_r,
  .explorer_classes = list(),
  .selectable_explorer_classes_r = shiny::reactive(character()),
  .addable_explorer_classes_r = .selectable_explorer_classes_r,
  .visible_explorer_classes_r = .selectable_explorer_classes_r,
  .label_list = label_explorer_selector(), ui = c("default",
  "minimal"), mode = c("modal", "in_place"),
  action_button_fun = shiny::actionButton)

Arguments

id
Module id.

input, output, session
Called by callModule.

.values
The .values list.

.root_node_r
A reactive returning an object of class ExplorerNode. This needs not necessarily to be the root node of an object of class ExplorerTree.

.explorer_classes
A list of objects of class ExplorerClass.

.selectable_explorer_classes_r
A reactive returning a character vector containing the ids of selectable explorer_classes. If character(), only group nodes are selectible (if .group_nodes_selectable = TRUE)

.addable_explorer_classes_r
A reactive returning a character vector containing the ids of explorer classes that are addable to the explorer.

.visible_explorer_classes_r
A reactive returning a character vector containing the ids of explorer classes that are displayed to the user. Group nodes are always displayed.

.label_list
A list created with label_explorer containing labels for all buttons used inside the explorer module.

ui
If "modal_default", the UI consists of an actionButton for selecting an element and an explorer, in which the selected element is shown. If "modal_minimal", only an actionButton showing the name of the selected element as label is shown. If "in_place" the selection is done in place.

action_button_fun
A function returning an HTML button, which is connected with shiny.

Value

The explorer_selector module returns a list containing the following reactives:

selected_node_r	An object of class ExplorerNode, which has been selected by the user. If no node is selected, the computation is stopped by req.

selected_node_or_null_r	An object of class ExplorerNode, which has been selected by the user or NULL, if no node is selected.