ExplorerNode

R6Class representing a node in an explorer tree which can be traversed using the explorer module. Each node has an explorer_class attached representing the general behaviour of nodes with this explorer_class in the explorer. Furthermore an object is attached representing the special behaviour and data associated with this particular node.

Important

You usually don't want to call explicitly ExplorerNode$new. Instead you should create a new explorer tree with ExplorerTree, take its root node with root <- explorer_tree$get_root_node() and afterwards call root$add_child().

Methods

new(id = NULL, node_storage = NULL, parent = NULL, is_group_node = TRUE, explorer_class = NULL, object = NULL)

Initialize a new node object.

id	Unique identifier of the node. If NULL, this identifier is created internally.

node_storage	Node storage of an object of class ExplorerTree. This argument is the reason, you should not call ExplorerNode$new explicitly.

parent	The parent node of this node, or NULL, if node is root node.

is_group_node	Logical indicating whether this node is a group node or not. Only group nodes can have child nodes.

explorer_class	Object of class ExplorerClass, which defines the behaviour of this node in the explorer.

object	An arbitrary object for storing information about the node.

add_child(id = NULL, is_group_node = TRUE, explorer_class = NULL, object = NULL, return = c("self", "child")

Initialize a new node object which is attached to the current node object as a child, but only if this node is a group node.

id	Unique identifier of the node. If NULL, this identifier is created internally. is_group_node

Logical indicating, whether the child node will be a group node or not.	explorer_class

Object of class ExplorerClass, which defines the behaviour of the child node in the explorer.	object

An arbitrary object for storing information about the child node.	return

children()

Get a list of all child nodes of this node object.

get_child(id)

Get the child node object with id == id.

get_child_ids()

Get the ids of all child nodes of this node object.

get_explorer_class()

Get the explorer_class associated with this node object.

get_id()

Get the id of the node object.

get_nth_child(n)

Get the nth child of this node object. The first added child is returned for coden = 1.

get_object()

Get the object associated with this node object.

is_group_node()

Returns a logical indicating whether this node is a group node or not.

remove_child(id)

Remove the child with id == id.

set_explorer_class(explorer_class)

Set the explorer_class associated with this node object.

set_object(object)

Set an arbitrary object associated with this node object.